feat: 스케쥴러 및 레디스 로직 수정

+++ src/main/java/cmf/commitField/domain/redpanda/RedpandaProducer.java
@@ -20,7 +20,7 @@ public class RedpandaProducer {
     }
 
     // 커밋 업데이트 전송 메서드
-    public void sendCommitUpdate(String username, int commitCount) {
+    public void sendCommitUpdate(String username, long commitCount) {
         String message = String.format("{\"user\": \"%s\", \"commits\": %d}", username, commitCount);
         kafkaTemplate.send(TOPIC, message);
         System.out.println("📨 Sent commit update to Redpanda: " + message);
